Build artifacts for Fedora 28
Also, get rid of restriction for branches and tags and instead rely on
manual action (using "when:manual"). This will allow us to re-build RPMs
on demand for future Fedora versions without having to introduce some
change that bumps the version, or create a special MR just to build
the RPMs.