Change URL handling in the MySQL parser
The "Text Form of Oracle CPU" page no longer contains full URL to the
main CPU page, the host part is now omitted.  This caused
advisory-parser to return incomplete URLs in from_url.  See
mprpic/advisory-parser#15 for more details.

Rather than doing a minimal fix of pre-pending 'https://www.oracle.com'
to the URL extracted form the page, this commit rewrites the way URLs
are processed in the MySQL advisory parser.  As CPU URLs do not contain
the unpredictable numeric id any more, it is no longer required to parse
the main CPU URL out of the text of the Text Form CPU page at all.  It
is now possible to easily translate the main page URL to the verbose
page URL and vice versa.  Therefore, this commit makes MySQL parser
accept either of the two URLs, and translates them as needed.