Update MySQL parser test
This adjusts the test for the MySQL parser to use the new CPU page URL
without the unpredictable numeric id.  The Apr 2019 CPU page used in the
test is no longer available at:

https://www.oracle.com/technetwork/security-advisory/cpuapr2019verbose-5072824.html

and is now available at:

https://www.oracle.com/security-alerts/cpuapr2019.html

The page as stored in the test_data directory was re-downloaded in the
new format.

Note that there is a problem in that CPU page related to CVE anchor
hrefs.  They are listed as:

<a id="CVE-2018-0734"></a>CVE-2018-0734

rather than:

<a id="CVE-2018-0734">CVE-2018-0734</a>

The latter is the format that was used before, but also what can be seen
in the current Apr 2020 CPU.  Presumably, the former link format is a
bug introduced when converting old CPU pages to the new format.  To
account for both formats, MySQL parser does not extract the CVE from the
<a> tag's string, but from the id attribute.

This commit also silences one test suite warning about invalid escape
sequence by properly marking regular expression as raw string.